dock1/requirements.txt
psycopg2
uwsgi
Flask
python-memcached